bootc: Cleanup running containers when process is manually killed
Before this change, if Ctrl+c was sent to kill the podman-bootc process
while the bootc install container was running, then the bootc install
container would continue to run in the background. This changes the
bootc struct to a singleton so the install bootc install container ID
can be stored on the object and used during cleanup.

Signed-off-by: ckyrouac <ckyrouac@redhat.com>