vm: Split linux and mac code into separate structs
Linux will use libvirt and mac will continue to use qemu. The common
bits are shared in the BootcVM struct and the common creation logic
is shared in cmd/run.go.

Using the libvirt library should be more stable since we will control
the version and we don't depend on the version of qemu of podman or the
version installed on the user's machine.

The SSH connection check was cleaned up in the process to wait until the
connection is completely available instead of assuming it is available
when the pid file exists and the port is open.

Signed-off-by: ckyrouac <ckyrouac@redhat.com>