vm: Fix rm when libvirt domain is already deleted
Prior to this, the call to rm would fail when the libvirt domain was
missing. This changes the bootcvm constructor to allow the domain lookup
to fail with a warning message. The constructor setup needs to be
furthur refined in the future to require the minimal set of static
parameters and load the VM creation (run) parameters only when running
the VM.

Signed-off-by: ckyrouac <ckyrouac@redhat.com>