tests: Create test framework for VM code
This modifies some of the VM code to enable testing it and
creates a few simple tests to demonstrate it works.

In order to keep the test code separate from production code, a fairly
significant refactor was required.

Primarily, almost all of the config code was moved into a new User struct.
This allows the user to be easily overridden in the test code.
I think it also better encapsulates all the paths related to the user config/caches.

This also cleans up the VM parameters by splitting the parameters into
two structs, NewVMParameters and RunVMParameters. This was also done in
an effort to make the code more testable, e.g. easily passing the
libvirtUri.

Signed-off-by: ckyrouac <ckyrouac@redhat.com>