Remove warning message
When removing a stopped VM the rm command shows a warning,
because the domain is already undefined. Let's remove that
warning since is the expected behavior, and only ignore
the libvirt's `ERR_NO_DOMAIN` error.

This is not ideal, since undefined domain is not expected
on stop, but still better than ignoring all errors.

Signed-off-by: German Maglione <gmaglione@redhat.com>