vm: Print console to stdout when booting VM
The console output streams to stdout while the VM is booting. When ssh
becomes available, the output stops streaming so the SSH connection can
be made.

On linux, the libvirt socket is used to stream the output.
On mac, a unix socket is created by qemu and connected to
using net.Dial.

Fixes #16

Signed-off-by: ckyrouac <ckyrouac@redhat.com>