podman: Use podman inspect for v4 path
When falling back to podman v4 to get the podman machine info,
instead of hard coding the values this uses the output of podman list
and podman inspect to verify a default machine is running and it is rootful.
This allows for multiple podman machines to be running by checking for
the machine flagged as default. When a default podman machine running as
rootful is not found, this prints a large banner style error
message informing the user that a rootful podman machine is required.

Signed-off-by: ckyrouac <ckyrouac@redhat.com>