libvirt: Remove some qemu:arg items from the template
This is a follow-up to e6268dc.
`-accel kvm` was apparently non-functional before e6268dc, and is no
longer useful since the libvirt domain definition explicitly requires
kvm.
`-cpu host` should be similar to `<cpu host="host-model">`, and the
impact on performance should not be as huge as switching from qemu to
kvm anyway.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>