Lock the cache dir for doing operations
This first implementation is far from ideal. To make it better
requires refactoring the code, since the cache logic is all over
the place. In any case, I think this is better than nothing.

I do not use 'github.com/containers/storage/pkg/lockfile' because
it does not allow to acquire an exclusive lock and after releasing
it, to acquire a shared one. After refactoring I'll try it again.

Signed-off-by: German Maglione <gmaglione@redhat.com>