disk: Don't pass stdin to install run, also set up cancellation
This fixes a regression in the previous commit; the attachment
takes over stdin and was competing with the ssh client.  We
have no need for interactivity, so don't pass stdin.

While we're here, also set up cancellation for the attachment
so that we know the operation is done when exiting the function,
as we also want to transfer ownership of stdout/stderr back.

Signed-off-by: Colin Walters <walters@verbum.org>