Add a storage object
This is the first part of a bigger refactoring. I tried to
make it as non-intrusive as possible, so it's not ideal.
But it will simplify the review, since this should have the
same behavior/mistakes than the previous version.

The current architecture of having all the actions as methods
of a VM object, and storing the config of the disk image, VM,
and the running Vm in s single json, is very inflexible.
In addition, the cache logic is all over the place, so let's
start with a simple cache implementation, that will allow us
to move to the final goal.

Signed-off-by: German Maglione <gmaglione@redhat.com>