Add function to inspect and pull the image
To use BIB we will need to also inpect images and pull them
if required. So, let's make a function to take care of this task
instead of copy-paste the same code.

I think we should go back to have a podman pkg, it will make it
easier in the future to switch between remote and local podman.

Signed-off-by: German Maglione <gmaglione@redhat.com>