Add a VM monitor for macOS
In order to switch from QEMU to krunkit we need to use gvproxy
to provide network support to the VM. We need to bound gvproxy
lifetime to krunkit lifetime, so there is not an "orphan"
daemon after krunkit finish.

We are using an "internal" command to monitor, in the future
this should be a different binary, but right now is the
easiest way. We save the pid fo krunkit instead of the monitor
because the `stop` command sends a SIGINT to that pid, and
we currently are capturing the signal at podman-bootc.

Signed-off-by: German Maglione <gmaglione@redhat.com>