Drop losetup wrapper
It was reported that `podman-bootc run quay.io/fedora/fedora-silverblue:41`
doesn't work because there `/usr/local` is a symlink, and our bind
mount hack breaks.

This was always intended to only be a temporary workaround
until the Linux kernel was fixed, and recent podman-machine (Fedora)
versions that appears to be true.

So just drop the hack.

(Now, `podman-bootc run <silverblue>` stil ldoesn't actually work by
 default because sshd doesn't run by default there, but that's
 unrelated to this tool)

Signed-off-by: Colin Walters <walters@verbum.org>