Build vm image for bootc installation VM
The vm image contains the virtualization stack to launch the virtual
machine, the files to compose the VM configuration in the bootc-data
volume.

The entrypoint prepares the bootc-data volume in order to be able to
boot from the bootc image mounted at that location plus the
configurations required by the installation.

The systemd services mount the virtiofs targets, while the
podman-vsock-proxy starts the proxy from VSOCK to the local unix socket
for podman.

The virtiofs-wrapper script is required in order to add extra option
when virtiofs is launched by libvirt. This is necessary in order to
correctly launch virtiofs inside an unprivileged container.

Signed-off-by: Alice Frosi <afrosi@redhat.com>