utils: add generic function for pointers
Go doesn't offer a generic method to convert a type in to its pointer.
This method is practical for the podman bindings where there are several
field which are pointer, and without this method, you would need to
define a variable and then pass its address. E.g for booleans:
   foo := utils.Ptr(true)
instead of
   t := true
   foo := &t

Signed-off-by: Alice Frosi <afrosi@redhat.com>